pyproject.toml
[tool.poetry]
name = "maakansha_assets"
version = "0.1.0"
description = "HR Onboarding Multi-Agent System"
authors = ["Maakansha"]
packages = [
    { include = "src" },
    { include = "agent.py" }
]

[tool.poetry.dependencies]
python = "^3.11"
google-adk = ">=1.23.0"
deprecated = "^1.2.14" 
google-cloud-aiplatform = "^1.38.0"
google-cloud-firestore = "^2.11.1"
google-cloud-storage = "^2.10.0"
google-cloud-bigquery = "^3.11.0"
google-api-python-client = "^2.88.0"
google-auth = "^2.19.0"
fpdf2 = "^2.7.0"
pydantic = "^2.0.0"
python-dateutil = "^2.8.2"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
absl-py = "^2.0.0"
tenacity = "^9.0.0"
google-cloud-bigquery-storage = "^2.24.0"
pandas = ">=2.0.0,<3.0.0"
db-dtypes = "^1.5.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
